% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_functions.R
\name{calc_entropy}
\alias{calc_entropy}
\title{Calculate the entropy for scRNA-seq Data}
\usage{
calc_entropy(counts)
}
\arguments{
\item{counts}{The counts matrix}
}
\value{
The entropy per gene
}
\description{
This will look at a matrix of counts data from an scRNA-seq experiment and
calculate the entropy of each gene.
}
\examples{

mat <- matrix(1:12, ncol = 3)
calc_entropy()

}
