% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_functions.R
\name{cut_entropy}
\alias{cut_entropy}
\title{Threshold Genes Based on Their Entropy}
\usage{
cut_entropy(counts, threshold)
}
\arguments{
\item{counts}{Counts matrix}

\item{threshold}{The threshold}
}
\value{
A vector
}
\description{
This is a description
}
